2021/01/31 전성원 1. 타이틀 바 없앰 2. 스플래쉬 화면 수정 (첫화면) 3. 어플 아이콘 수정
